+++ modules/apps/foundation/frontend-js/frontend-js-web/package.json
@@ -16,7 +16,6 @@
 		}
 	},
 	"dependencies": {
-		"check-source-formatting": "^2.0.4",
 		"clay": "2.0.0-rc.5",
 		"liferay-amd-loader": "3.1.1",
 		"metal": "^2.16.2",
@@ -30,6 +29,7 @@
 	"devDependencies": {
 		"babel-loader": "^7.1.2",
 		"babel-preset-env": "^1.6.0",
+		"check-source-formatting": "^2.0.4",
 		"jest": "^20.0.4",
 		"liferay-jest-junit-reporter": "0.0.3",
 		"liferay-module-config-generator": "^1.2.1",
@@ -56,8 +56,6 @@
 		"checkFormat": "npm run csf",
 		"csf": "csf src/**/liferay/**/*.js test/**/*.js",
 		"format": "npm run csf -- -i",
-		"lint": "eslint src/**/*.es.js test/**/*.es.js",
-		"prettier": "prettier-eslint src/**/*.es.js test/**/*.es.js",
 		"start": "webpack-dev-server --config webpack.dev.js",
 		"test": "jest",
 		"webpack": "webpack --config webpack.prod.js",
