+++ modules/apps/foundation/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/global.es.js
@@ -1,0 +1,5 @@
+'use strict';
+
+import register from './portlet_hub/register.es';
+
+export {register};
