--- modules/apps/foundation/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/global.js
@@ -1,5 +1,0 @@
-'use strict';
-
-import register from './portlet_hub/register';
-
-export {register};
