+++ modules/apps/foundation/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/portlet_hub/PortletInit.es.js
@@ -6,12 +6,17 @@ const portletRegex = '^portlet[.].*';
 
 /**
  * PortletInit
+ * @review
  */
+
 class PortletInit {
+
 	/**
 	 * Constructor for PortletInit
 	 * @param {string} portletId
+	 * @review
 	 */
+
 	constructor(portletId) {
 		this._portletId = portletId;
 	}
@@ -20,18 +25,20 @@ class PortletInit {
 	 * Initiates a portlet action using the specified action parameters and
 	 * element arguments.
 	 *
+	 * @memberof PortletInit
 	 * @param {PortletParameters} params Action parameters to be added to the
 	 * URL
 	 * @param {HTMLFormElement} element DOM element of form to be submitted
 	 * @returns {Promise} A Promise object that is resolved with no argument
 	 * when the action request has completed.
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
 	 * @throws {AccessDeniedException} Thrown if a blocking operation is
 	 * already in progress.
 	 * @throws {NotInitializedException} Thrown if a portlet ID is provided,
 	 * but no onStateChange listener has been registered.
-	 * @memberof PortletInit
 	 */
+
 	action(params, element) {
 		throw new Error('"action" is not yet implemented');
 	}
@@ -39,12 +46,14 @@ class PortletInit {
 	/**
 	 * Adds a listener function for specified event type.
 	 *
+	 * @memberof PortletInit
 	 * @param {string} type The type of listener
 	 * @param {function} handler Function called when event occurs
 	 * @returns {object} A handle that can be used to remove the event listener
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
-	 * @memberof PortletInit
 	 */
+
 	addEventListener(type, handler) {
 		if (arguments.length > 2) {
 			throw new TypeError(
@@ -71,6 +80,7 @@ class PortletInit {
 	 * for the page state according to the resource parameters, cacheability
 	 * option, and resource ID provided.
 	 *
+	 * @memberof   PortletInit
 	 * @param {PortletParameters} params Resource parameters to be added to
 	 * the URL
 	 * @param {string} cache Cacheability option. The strings defined under
@@ -79,9 +89,10 @@ class PortletInit {
 	 * @returns {Promise} A Promise object. Returns a string representing the
 	 * resource URL on successful resolution. Returns an Error object containing
 	 * a descriptive message on failure.
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
-	 * @memberof   PortletInit
 	 */
+
 	createResourceUrl(params, cache, resourceId) {
 		throw new Error('"createResourceUrl" is not yet implemented');
 	}
@@ -89,12 +100,14 @@ class PortletInit {
 	/**
 	 * Dispatches a client event.
 	 *
+	 * @memberof PortletInit
 	 * @param {string} type The type of listener
 	 * @param {any} payload The payload to be delivered
 	 * @returns {number} The number of events queued for delivery
+	 * @review
 	 * @throws {TypeError} Thrown if the type is a system event type
-	 * @memberof PortletInit
 	 */
+
 	dispatchClientEvent(type, payload) {
 		if (arguments.length > 2) {
 			throw new TypeError(
@@ -110,23 +123,28 @@ class PortletInit {
 			throw new TypeError('The event type is invalid: ' + type);
 		}
 
-		return PortletInit._clientEventListeners.reduce((amount, listener) => {
-			if (type.match(listener.type)) {
-				listener.handler(payload);
+		return PortletInit._clientEventListeners.reduce(
+			(amount, listener) => {
+				if (type.match(listener.type)) {
+					listener.handler(payload);
 
-				amount++;
-			}
+					amount++;
+				}
 
-			return amount;
-		}, 0);
+				return amount;
+			},
+			0
+		);
 	}
 
 	/**
 	 * Tests whether a blocking operation is in progress.
 	 *
-	 * @returns {boolean} true if a blocking
 	 * @memberof PortletInit
+	 * @returns {boolean} true if a blocking
+	 * @review
 	 */
+
 	isInProgress() {
 		return PortletInit._busy;
 	}
@@ -134,11 +152,13 @@ class PortletInit {
 	/**
 	 * Creates and returns a new PortletParameters object.
 	 *
+	 * @memberof PortletInit
 	 * @param {PortletParameters} param An optional PortletParameters object to
 	 * be copied
 	 * @returns {PortletParameters} The new parameters object
-	 * @memberof PortletInit
+	 * @review
 	 */
+
 	newParameters(param) {
 		throw new Error('"newParameters" is not yet implemented');
 	}
@@ -146,10 +166,12 @@ class PortletInit {
 	/**
 	 * Creates and returns a new RenderState object.
 	 *
+	 * @memberof PortletInit
 	 * @param {RenderState} state An optional RenderState object to be copied
 	 * @returns {RenderState} The new RenderState object
-	 * @memberof PortletInit
+	 * @review
 	 */
+
 	newState(state) {
 		throw new Error('"newState" is not yet implemented');
 	}
@@ -159,12 +181,14 @@ class PortletInit {
 	 * The handle must be the same object previously returned by the
 	 * addEventListener function.
 	 *
+	 * @memberof PortletInit
 	 * @param {object} handle The handle of the listener to be removed
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
 	 * @throws {AccessDeniedException} Thrown if the event listener associated
 	 * with this handle was registered by a different portlet
-	 * @memberof PortletInit
 	 */
+
 	removeEventListener(handle) {
 		if (arguments.length > 1) {
 			throw new TypeError(
@@ -180,6 +204,7 @@ class PortletInit {
 
 		// Currently only checking for client events, eventually system events
 		// will need to be checked as well.
+
 		const index = PortletInit._clientEventListeners.indexOf(handle);
 
 		if (index !== -1) {
@@ -187,7 +212,7 @@ class PortletInit {
 		}
 		else {
 			throw new TypeError(
-				`The event listener handle doesn't match any listeners.`
+				'The event listener handle doesn\'t match any listeners.'
 			);
 		}
 	}
@@ -196,14 +221,16 @@ class PortletInit {
 	 * Sets the render state, which consists of the public and private render
 	 * parameters, the portlet mode, and the window state.
 	 *
+	 * @memberof PortletInit
 	 * @param {RenderState} state The new state to be set
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
 	 * @throws {AccessDeniedException} Thrown if a blocking operation is
 	 * already in progress.
 	 * @throws {NotInitializedException} Thrown if a portlet ID is provided, but
 	 * no onStateChange listener has been registered.
-	 * @memberof PortletInit
 	 */
+
 	setRenderState(state) {
 		throw new Error('"setRenderState" is not yet implemented');
 	}
@@ -212,19 +239,21 @@ class PortletInit {
 	 * Starts partial action processing and returns a
 	 * {@link PartialActionInit} object to the caller.
 	 *
+	 * @memberof PortletInit
 	 * @param {PortletParameters} params Action parameters to be added to the
 	 * URL
 	 * @returns {Promise} A Promise object. Returns a {PortletActionInit} object
 	 * containing a partial action URL and the setPageState callback function on
 	 * successful resolution. Returns an Error object containing a descriptive
 	 * message on failure.
+	 * @review
 	 * @throws {TypeError} Thrown if the input parameters are invalid
 	 * @throws {AccessDeniedException} Thrown if a blocking operation is already
 	 * in progress.
 	 * @throws {NotInitializedException} Thrown if a portlet ID is provided, but
 	 * no onStateChange listener has been registered.
-	 * @memberof PortletInit
 	 */
+
 	startPartialAction(params) {
 		throw new Error('"startPartialAction" is not yet implemented');
 	}
@@ -235,11 +264,13 @@ class PortletInit {
 	 * @param {string} type The type of listener
 	 * @param {function} handler Function called when event occurs
 	 * @returns {object} A handle that can be used to remove the event listener
+	 * @review
 	 */
+
 	_addClientEventListener(type, handler) {
 		const listener = {
 			handler,
-			type,
+			type
 		};
 
 		PortletInit._clientEventListeners.push(listener);
@@ -251,19 +282,23 @@ class PortletInit {
 /**
  * Determines if blocking action is currently in process.
  *
- * @type {boolean}
  * @memberof PortletInit
+ * @review
  * @static
+ * @type {boolean}
  */
+
 PortletInit._busy = false;
 
 /**
  * Contains client event listeners added from all instances of PortletInit.
  *
- * @type {array}
  * @memberof PortletInit
+ * @review
  * @static
+ * @type {array}
  */
+
 PortletInit._clientEventListeners = [];
 
 export default PortletInit;
