+++ modules/apps/foundation/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/portlet_hub/register.es.js
@@ -2,7 +2,7 @@
 
 import Promise from 'metal-promise';
 
-import PortletInit from './PortletInit';
+import PortletInit from './PortletInit.es';
 
 /**
  * Registers a portlet client with the portlet hub.
@@ -11,7 +11,9 @@ import PortletInit from './PortletInit';
  * @returns {Promise} A Promise object. Returns an {@link PortletInit} object
  * containing functions for use by the portlet client on successful resolution.
  * Returns an Error object containing a descriptive message on failure.
+ * @review
  */
+
 function register(portletId) {
 	return new Promise(
 		(resolve, reject) => {
@@ -23,7 +25,7 @@ function register(portletId) {
 			}
 		}
 	);
-};
+}
 
 export {register};
 export default register;
